requirements.txt
"""
Gereksinimler (Requirements)
============================

Bu dosyayı kullanarak tüm bağımlılıkları tek seferde kurabilirsiniz:
    pip install -r requirements.txt
"""

requests>=2.32.0
beautifulsoup4>=4.13.0
tkcalendar>=1.6.0
python-dotenv>=1.0.0
pystray>=0.19.0
Pillow>=10.0.0
